Fictional Report Card Data - Secondary

Fictional data for Secondary Report Card Marks

Usage

sec_rc

Format

A dataframe with 200 rows and 13 variables

id

Unique student identifier

school

School where student is enrolled

course_number

Course Code (i.e. ENG2D, SNC1P)

potential_credit_hrs

Number of potential credits the student can earn from that course

earned_credit_hrs

Number of credits the student earned from that course

mid__term

Mid term mark

final

Final mark

collaboration

Learning Skill - Collaboration

independent_work

Learning Skill - Independent Work

initiative

Learning Skill - Initiative

organization

Learning Skill - Organization

responsibility

Learning Skill - Responsibility

self_regulation

Learning Skill - Self Regulation